ASoC: wm_adsp: fix a type cast warning in debug output
(cherry picked from commit d60bf2c001ea7a7928643b11f4e483cf83e3046e)
Reviewed-on: http://gerrit.mot.com/679259
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>

+++ sound/soc/codecs/wm_adsp.c
@@ -2907,7 +2907,7 @@ int wm_adsp_stream_read(struct wm_adsp *adsp, char __user *buf, size_t count)
 		count = avail;
 
 	adsp_dbg(adsp, "%s: avail=%d toend=%d count=%d\n",
-		 __func__, avail, to_end, count);
+		 __func__, avail, to_end, (int)count);
 
 	if (count > to_end) {
 		if (copy_to_user(buf,
